backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-multipart==0.0.9
scikit-learn==1.5.2
numpy==2.1.1
PyJWT==2.9.0
reportlab==4.2.2
psycopg2-binary==2.9.9
opencv-python==4.8.1.78
mediapipe==0.10.5
ultralytics==8.3.5
